+++ Programs/usb_libusb.c
@@ -148,7 +148,7 @@ usbSubmitRequest (
   UsbDevice *device,
   unsigned char endpointAddress,
   void *buffer,
-  int length,
+  size_t length,
   void *data
 ) {
   errno = ENOSYS;
@@ -182,7 +182,7 @@ usbReadEndpoint (
   UsbDevice *device,
   unsigned char endpointNumber,
   void *buffer,
-  unsigned int length,
+  size_t length,
   int timeout
 ) {
   UsbDeviceExtension *devx = device->extension;
@@ -237,7 +237,7 @@ usbWriteEndpoint (
   UsbDevice *device,
   unsigned char endpointNumber,
   const void *buffer,
-  unsigned int length,
+  size_t length,
   int timeout
 ) {
   UsbDeviceExtension *devx = device->extension;
