+++ Programs/usb_linux.c
@@ -390,7 +390,7 @@ usbSubmitRequest (
   UsbDevice *device,
   unsigned char endpointAddress,
   void *buffer,
-  int length,
+  size_t length,
   void *context
 ) {
   UsbDeviceExtension *devx = device->extension;
@@ -553,7 +553,7 @@ static int
 usbBulkTransfer (
   UsbEndpoint *endpoint,
   void *buffer,
-  unsigned int length,
+  size_t length,
   int timeout
 ) {
   UsbDeviceExtension *devx = endpoint->device->extension;
@@ -624,7 +624,7 @@ usbReadEndpoint (
   UsbDevice *device,
   unsigned char endpointNumber,
   void *buffer,
-  unsigned int length,
+  size_t length,
   int timeout
 ) {
   int count = -1;
@@ -671,7 +671,7 @@ usbWriteEndpoint (
   UsbDevice *device,
   unsigned char endpointNumber,
   const void *buffer,
-  unsigned int length,
+  size_t length,
   int timeout
 ) {
   UsbEndpoint *endpoint;
